Refactored shooting code on all objects so that a tower is removed from the array list once destroyed

+++ Assignment/ObjectMenu.pde
@@ -15,7 +15,7 @@ class ObjectMenu
    
     void objMenu()
   {
-    float menuLane=width-objectW;
+    float menuLane=(float)width-(float)objectW;
     
     Tank mTank=new Tank();
     mTank.pos=new PVector(menuLane,objectW);
